Page header (image)
Now that we even have a tag cloud, I think it's time we made this look a little more like a blog site with a header.

Tweaked scrollToItem call to avoid scrolling to first item on first page load, which also allowed me to address that annoying line of code.